Refactor to group grades-senders in its own submodule/folder
